fix: Change type of Grid and Cell children
