--- packages/ui/src/components/Grid/types.d.ts
@@ -1,17 +1,0 @@
-import { ReactElement } from 'react';
-
-type Children = ReactElement[] | ReactElement;
-type Size = number;
-type Offset =
-  | {
-  before: Size;
-  after?: Size;
-}
-  | {
-  before?: Size;
-  after: Size;
-}
-  | {
-  before: Size;
-  after: Size;
-};
